requirement.txt
fastapi
uvicorn
pydantic
langchain
python-dotenv
pandas
langchain-community
langsmith
pymongo
